pyproject.toml
[project]
name = "telnyx-fax-oss-example"
version = "0.1.0"
description = "API-only Telnyx fax core with upload, send, webhooks, status, and cancel"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.2.1",
  "boto3>=1.35.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "phonenumbers>=8.13.40",
  "pypdf>=4.3.0",
  "Pillow>=10.4.0",
  "email-validator>=2.2.0",
  "cryptography>=43.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.5",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
